+++ src/components/bbn-resource-calculator/bbn-resource-calculator.tsx
