+++ src/components/gauge/index.tsx
